refactor(spec): move failure summarization into CLI scan layer
summarizeExtractionFailures and its FailureSample/ExtractionFailureReport
types only ever fed the scan CLI path, so they don't belong in the shared
extractor. Move them next to decideScanOutcome in the CLI command; the
extractor now returns raw extraction results only. Tests relocated to
tests/cli/spec-scan.test.ts. No behavior change.
