+++ packages/spec-consolidator/src/index.ts
@@ -56,16 +56,8 @@ export {
 } from './prompt.js';
 export type { LlmExtraction, LlmClaim } from './prompt.js';
 
-export {
-  extractClaims,
-  summarizeExtractionFailures,
-} from './extractor.js';
-export type {
-  ExtractOptions,
-  ExtractResult,
-  ExtractionFailureReport,
-  FailureSample,
-} from './extractor.js';
+export { extractClaims } from './extractor.js';
+export type { ExtractOptions, ExtractResult } from './extractor.js';
 
 export { mergeClaims, candidateFingerprint } from './merger.js';
 export type { MergeResult, DecidedConflict } from './merger.js';
