+++ tests/cli/spec-scan.test.ts
@@ -1,7 +1,10 @@
 import { describe, it, expect } from 'vitest';
-import { decideScanOutcome } from '../../tools/cli/src/commands/spec';
+import {
+  decideScanOutcome,
+  summarizeExtractionFailures,
+  type ExtractionFailureReport,
+} from '../../tools/cli/src/commands/spec';
 import { describeClaudePreflightFailure } from '../../tools/cli/src/lib/claude-preflight';
-import type { ExtractionFailureReport } from '@truecourse/spec-consolidator';
 
 /**
  * Policy for how `truecourse spec scan` ends (issue #474): a total extraction
@@ -59,6 +62,60 @@ describe('decideScanOutcome', () => {
     });
     expect(outcome.exitCode).toBe(0);
     expect(outcome.outro).toBe('4 open.');
+  });
+});
+
+/**
+ * `summarizeExtractionFailures` collapses duplicate per-block errors into
+ * counted samples and flags a total wipeout. It feeds `decideScanOutcome`, so
+ * it lives in the CLI scan layer alongside it (not the shared extractor).
+ */
+describe('summarizeExtractionFailures', () => {
+  const fail = (error: string, n = 1) => Array.from({ length: n }, () => ({ error }));
+
+  it('reports nothing when there are no failures', () => {
+    const r = summarizeExtractionFailures({ failures: [], blocksAttempted: 10 });
+    expect(r).toEqual({ total: 0, allFailed: false, samples: [] });
+  });
+
+  it('flags allFailed only when every attempted block failed', () => {
+    expect(
+      summarizeExtractionFailures({ failures: fail('boom', 10), blocksAttempted: 10 }).allFailed,
+    ).toBe(true);
+    expect(
+      summarizeExtractionFailures({ failures: fail('boom', 4), blocksAttempted: 10 }).allFailed,
+    ).toBe(false);
+  });
+
+  it('never flags allFailed when no blocks were attempted', () => {
+    const r = summarizeExtractionFailures({ failures: [], blocksAttempted: 0 });
+    expect(r.allFailed).toBe(false);
+  });
+
+  it('collapses duplicate messages into samples with counts, most frequent first', () => {
+    const r = summarizeExtractionFailures({
+      failures: [...fail('A', 5), ...fail('B', 2), ...fail('C', 1)],
+      blocksAttempted: 8,
+    });
+    expect(r.total).toBe(8);
+    expect(r.samples).toEqual([
+      { message: 'A', count: 5 },
+      { message: 'B', count: 2 },
+      { message: 'C', count: 1 },
+    ]);
+  });
+
+  it('caps samples at the limit (default 3, overridable)', () => {
+    const failures = [
+      ...fail('A', 4),
+      ...fail('B', 3),
+      ...fail('C', 2),
+      ...fail('D', 1),
+    ];
+    expect(summarizeExtractionFailures({ failures, blocksAttempted: 10 }).samples).toHaveLength(3);
+    expect(
+      summarizeExtractionFailures({ failures, blocksAttempted: 10 }, { sampleLimit: 2 }).samples,
+    ).toHaveLength(2);
   });
 });
